+++ src/app/[locale]/components/input-data/utils.tsx
