+++ src/app/[locale]/report/page.tsx
@@ -4,9 +4,8 @@ import { useState } from "react";
 import { useTranslations } from "next-intl";
 import { useSelector } from "react-redux";
 import { setProject } from "../redux/project-slice";
-import { setData } from "../redux/report-slice";
 import { ProjectState } from "../redux/store";
-import { Tab } from "../input-data/page";
+import { Tab } from "../components/input-data/input-data";
 import Summary from "./summary";
 import Graph from "./graph";
 import Detail from "./detail";
